backend/python_service/requirements.txt
telethon
fastapi
uvicorn
python-dotenv
